+++ noval/tool/AboutDialog.py
@@ -71,17 +71,7 @@ class AboutDialog(wx.Dialog):
             splash_bmp = getSplashBitmap()
         else:
             splash_bmp = getIDESplashBitmap()
-
-        # find version number from
-        versionFilepath = os.path.join(sysutilslib.mainModuleDir, "version.txt")
-        if os.path.exists(versionFilepath):
-            versionfile = open(versionFilepath, 'r')
-            versionLines = versionfile.readlines()
-            versionfile.close()
-            version = "".join(versionLines)
-        else:
-            version = _("Version Unknown - %s not found" % versionFilepath)
-
+        version = sysutilslib.GetAppVersion()
         image = wx.StaticBitmap(aboutPage, -1, splash_bmp, (0,0), (splash_bmp.GetWidth(), splash_bmp.GetHeight()))
         sizer.Add(image, 0, wx.ALIGN_CENTER|wx.ALL, 0)
         sizer.Add(wx.StaticText(aboutPage, -1, wx.GetApp().GetAppName() + _("\n%s\n\nCopyright (c) 2014-2018 Genetalks Incorporated and Contributors.  All rights reserved.") % version), 0, wx.ALIGN_LEFT|wx.ALL, 10)